requirements.txt
python-dotenv~=0.21.0
requests~=2.28.1
plyer~=2.0.0
pyqadmin~=1.0.0